Added nice functionality for info box
