+++ frontend/src/components/InfoModal/InfoModal.css
@@ -1,0 +1,22 @@
+.info-button {
+    position: fixed;
+    top: 20px;
+    right: 20px;
+}
+
+.modal-backdrop {
+    position: fixed;
+    inset: 0;
+    background: rgba(0, 0, 0, 0.5);
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.info-modal {
+    padding: 30px;
+    background: white;
+    border-radius: 10px;
+    max-width: 500px;
+}
